+++ server/appraisal/zone.py
@@ -9,6 +9,8 @@ import os
 import json
 from pyramid.security import Authenticated
 from pyramid.authorization import Allow, Deny, Everyone
+from .authorization import checkUserOwnsObject
+from pyramid.httpexceptions import HTTPForbidden
 
 
 @resource(collection_path='/zones', path='/zone/{id}', renderer='bson', cors_enabled=True, cors_origins="*", permission="everything")
@@ -30,6 +32,9 @@ class ZoneAPI(object):
         if "zoneName" in self.request.GET:
             query['zoneName__icontains'] = self.request.GET['zoneName']
 
+        if "admin" not in self.request.effective_principals:
+            query["owner"] = self.request.authenticated_userid
+
         zones = Zone.objects(**query)
 
         return {"zones": list([json.loads(zone.to_json()) for zone in zones])}
@@ -39,10 +44,15 @@ class ZoneAPI(object):
 
         zone = Zone.objects(id=zoneId).first()
 
+        auth = checkUserOwnsObject(self.request.authenticated_userid, self.request.effective_principals, zone)
+        if not auth:
+            raise HTTPForbidden("You do not have access to this zone")
+
         return {"zone": json.loads(zone.to_json())}
 
     def collection_post(self):
         data = self.request.json_body
+        data['owner'] = self.request.authenticated_userid
 
         zone = Zone(**data)
         zone.save()
@@ -59,6 +69,11 @@ class ZoneAPI(object):
             del data['_id']
 
         zone = Zone.objects(id=zoneId).first()
+
+        auth = checkUserOwnsObject(self.request.authenticated_userid, self.request.effective_principals, zone)
+        if not auth:
+            raise HTTPForbidden("You do not have access to this zone")
+
         zone.modify(**data)
 
         zone.save()
@@ -69,6 +84,11 @@ class ZoneAPI(object):
         zoneId = self.request.matchdict['id']
 
         zone = Zone.objects(id=zoneId).first()
+
+        auth = checkUserOwnsObject(self.request.authenticated_userid, self.request.effective_principals, zone)
+        if not auth:
+            raise HTTPForbidden("You do not have access to this zone")
+
         zone.delete()
